refactor: move Grid into SharedComponents
