+++ src/components/DownloadAndShareComponents/DownloadCardGroup/DownloadCardGroup.tsx
@@ -11,7 +11,7 @@ import { sortDownloadResources } from "../helpers/sortResources";
 
 import { LessonDownloadsData } from "@/node-lib/curriculum-api";
 import Box from "@/components/SharedComponents/Box";
-import Grid from "@/components/Grid";
+import Grid from "@/components/SharedComponents/Grid";
 
 export type DownloadCardGroupProps = {
   downloads?: LessonDownloadsData["downloads"];
