+++ src/components/Forms/NewsletterForm/NewsletterForm.stories.tsx
@@ -1,8 +1,9 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import Grid, { GridArea } from "../../Grid";
-
 import Component from ".";
+
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
+
 
 export default {
   component: Component,
