+++ src/components/Forms/NewsletterForm/NewsletterFormWrap.tsx
@@ -8,7 +8,7 @@ import { P } from "@/components/SharedComponents/Typography";
 import CardTitle from "@/components/SharedComponents/Card/CardComponents/CardTitle";
 import Card, { CardProps } from "@/components/SharedComponents/Card";
 import AnchorTarget from "@/components/SharedComponents/AnchorTarget";
-import Grid, { GridArea } from "@/components/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 
 /**
  * Newsletter Form wrapper adds copy around around NewsletterForm.
