+++ src/components/HomePageTabs/PupilTab/PupilTab.tsx
@@ -4,8 +4,7 @@ import ImageContainer from "../ImageContainer/ImageContainer";
 
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Flex from "@/components/SharedComponents/Flex";
-import { GridArea } from "@/components/Grid";
-import Grid from "@/components/Grid/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Svg from "@/components/SharedComponents/Svg";
 import Typography, { Heading } from "@/components/SharedComponents/Typography";
