+++ src/components/HomePageTabs/TeachersTab/TeachersTab.tsx
@@ -10,8 +10,7 @@ import useSearch from "@/context/Search/useSearch";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import ResourceSelectorCard from "@/components/ResourceSelectorCard/ResourceSelectorCard";
 import { KeyStageKeypadProps } from "@/components/KeyStageKeypad/KeyStageKeypad";
-import Grid from "@/components/Grid/Grid";
-import { GridArea } from "@/components/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import KeyStageKeypad from "@/components/KeyStageKeypad";
 
 type TeacherTabProps = {
