+++ src/components/KeyStageKeypad/KeyStageKeypad.tsx
@@ -7,9 +7,8 @@ import useAnalytics from "../../context/Analytics/useAnalytics";
 import type { KeyStageTitleValueType } from "../../browser-lib/avo/Avo";
 import useAnalyticsPageProps from "../../hooks/useAnalyticsPageProps";
 import useIsCurrent from "../MenuLinks/useIsCurrent";
-import Grid from "../Grid/Grid";
-import { GridArea } from "../Grid";
 
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import { Heading, UL, LI, P } from "@/components/SharedComponents/Typography";
 
 export type KeypadItem = TeachersHomePageData["keyStages"][number];
