+++ src/components/Lesson/LessonAppearsIn/LessonAppearsIn.tsx
@@ -1,5 +1,5 @@
 import Flex from "@/components/SharedComponents/Flex";
-import Grid, { GridArea } from "@/components/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import { LessonPathwayCard } from "@/components/Lesson/LessonPathwayCard/LessonPathwayCard";
 import { TagFunctional } from "@/components/TagFunctional";
 import { TagColor } from "@/components/TagFunctional/TagFunctional";
