+++ src/components/SharedComponents/Grid/Grid.tsx
@@ -1,7 +1,7 @@
 import styled, { css } from "styled-components";
 
-import { NullablePixelSpacing, PixelSpacing } from "../../styles/theme";
-import responsive, { ResponsiveValues } from "../../styles/utils/responsive";
+import { NullablePixelSpacing, PixelSpacing } from "../../../styles/theme";
+import responsive, { ResponsiveValues } from "../../../styles/utils/responsive";
 
 import Box, { BoxProps } from "@/components/SharedComponents/Box";
