+++ src/components/SharedComponents/Grid/GridArea.tsx
@@ -1,7 +1,7 @@
 import styled, { css } from "styled-components";
 
-import responsive, { ResponsiveValues } from "../../styles/utils/responsive";
-import { SpacingProps } from "../../styles/utils/spacing";
+import responsive, { ResponsiveValues } from "../../../styles/utils/responsive";
+import { SpacingProps } from "../../../styles/utils/spacing";
 
 import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
