+++ src/components/SharedComponents/Typography/LI.tsx
@@ -3,10 +3,10 @@ import styled, { css, CSSProperties } from "styled-components";
 import flex from "../../../styles/utils/flex";
 import responsive, { ResponsiveValues } from "../../../styles/utils/responsive";
 import typography, { TypographyProps } from "../../../styles/utils/typography";
-import { GridAreaProps } from "../../Grid";
-import { gridArea } from "../../Grid/GridArea";
 import { box } from "../Box";
 
+import { GridAreaProps } from "@/components/SharedComponents/Grid";
+import { gridArea } from "@/components/SharedComponents/Grid/GridArea";
 import { FlexProps } from "@/components/SharedComponents/Flex";
 
 type ListItemProps = {
