+++ src/components/SharedComponents/Typography/OL.tsx
@@ -2,9 +2,9 @@ import styled from "styled-components";
 
 import { margin, MarginProps } from "../../../styles/utils/spacing";
 import getFontFamily from "../../../styles/themeHelpers/getFontFamily";
-import { GridProps } from "../../Grid";
-import { grid } from "../../Grid/Grid";
 
+import { GridProps } from "@/components/SharedComponents/Grid";
+import { grid } from "@/components/SharedComponents/Grid/Grid";
 import { box, BoxProps } from "@/components/SharedComponents/Box";
 
 /**
