+++ src/components/SharedComponents/Typography/UL.tsx
@@ -1,8 +1,7 @@
 import styled, { css } from "styled-components";
 
-import { GridProps } from "../../Grid";
-import { grid } from "../../Grid/Grid";
-
+import { GridProps } from "@/components/SharedComponents/Grid";
+import { grid } from "@/components/SharedComponents/Grid/Grid";
 import { box, BoxProps } from "@/components/SharedComponents/Box";
 import { FlexProps } from "@/components/SharedComponents/Flex";
 import flex from "@/styles/utils/flex";
