+++ src/components/pages/CurriculumInfo/tabs/UnitsTab/UnitsTab.tsx
@@ -9,8 +9,7 @@ import Icon from "@/components/SharedComponents/Icon";
 import OutlineHeading from "@/components/OutlineHeading/OutlineHeading";
 import Button from "@/components/SharedComponents/Button/Button";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders/BrushBorders";
-import GridArea from "@/components/Grid/GridArea";
-import Grid from "@/components/Grid/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import Radio from "@/components/RadioButtons/Radio";
 import RadioGroup from "@/components/RadioButtons/RadioGroup";
 import Sidebar from "@/components/Sidebar/Sidebar";
