+++ src/pages/about-us/work-with-us.tsx
@@ -10,7 +10,7 @@ import AboutContactCard from "@/components/AboutContactCard";
 import Typography, { Heading } from "@/components/SharedComponents/Typography";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Flex from "@/components/SharedComponents/Flex";
-import Grid, { GridArea } from "@/components/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import AboutIntroCard from "@/components/AboutIntoCard/AboutIntroCard";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders";
 import AboutUsSummaryCard from "@/components/pages/AboutUs/AboutUsSummaryCard";
