+++ src/pages/develop-your-curriculum.tsx
@@ -16,8 +16,7 @@ import Box from "@/components/SharedComponents/Box";
 import BoxBorders from "@/components/SpriteSheet/BrushSvgs/BoxBorders";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import CardLink from "@/components/SharedComponents/Card/CardLink";
-import Grid from "@/components/Grid";
-import GridArea from "@/components/Grid/GridArea";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import Cover from "@/components/Cover";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders";
