+++ src/pages/teachers/early-release-units.tsx
@@ -8,7 +8,7 @@ import earlyReleaseExemplarUnitsFixture from "@/node-lib/curriculum-api/fixtures
 import { Heading, Hr, P } from "@/components/SharedComponents/Typography";
 import Box from "@/components/SharedComponents/Box";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders/BrushBorders";
-import Grid, { GridArea } from "@/components/Grid";
+import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import Card from "@/components/SharedComponents/Card";
 import Flex from "@/components/SharedComponents/Flex";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
